% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible.R
\name{credibleInterval}
\alias{credibleInterval}
\title{Generate credible intervals for model parameters}
\usage{
credibleInterval(mod.out, param, prob = 0.95)
}
\arguments{
\item{mod.out}{model output from running \code{simpleLM} or \code{modularLM}.}

\item{param}{(optional) A character vector of variable names for which
credible intervals are desired. If not specified, credible intervals for all
parameters are returned.}

\item{prob}{Specifies the high probability density region of the posterior
distribution. Default is 0.95, which corresponds to a 95\% credible interval.}
}
\value{
An object of class \code{credibleInterval} containing the lower and upper
bounds of the credible intervals for the specified parameters, along with
posterior medians.
}
\description{
Generate credible intervals for model parameters
}
\examples{
set.seed(1729)
n <- 100
beta0 <- c(5, -2, 1, 4)
sd0 <- 1
wts0 <- rep(1, n)

p <- length(beta0)
X <- cbind(rep(1, n), sapply(1:(p - 1), function(x) rnorm(n)))
mu <- X \%*\% beta0
y <- rnorm(n, mu, sd0 / sqrt(wts0))
dat <- cbind(y, X[, -1])
dat <- as.data.frame(dat)
names(dat) <- c("y", paste("x", 1:(p - 1), sep = ""))
nS <- 100
modular.x3 <- matrix(rnorm(length(dat$x3) * nS, mean = rep(dat$x3, nS), sd = 0.01),
                     nrow = length(dat$x3), ncol = nS)
mod1 <- modularLM(y ~ x1 + x2, post.var = list(x3 = modular.x3),
                  data = dat)
mod1.summary <- credibleInterval(mod1)
}
