% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularLM.R
\name{modularLM}
\alias{modularLM}
\title{Fit a modular Bayesian linear model}
\usage{
modularLM(formula, data, post.var, wts, priors, ...)
}
\arguments{
\item{formula}{a symbolic description of the non-modular component of the
regression model to be fit.}

\item{data}{An optional data frame containing the variables in the model.}

\item{post.var}{A list tagged with modular predictor names, each containing
a matrix of posterior samples for that modular predictor. In case of
interaction terms involving a modular predictor, the interaction term should
also be included as a separate entry in this list.}

\item{wts}{A numeric vector of weights for the observations.}

\item{priors}{A list containing prior specifications for model parameters.}

\item{...}{Additional arguments (not currently used).}
}
\value{
An object of class \code{modularLM} containing the fitted model and
posterior samples.
}
\description{
Fit a modular Bayesian linear model
}
\examples{
set.seed(1729)
n <- 100
beta0 <- c(5, -2, 1, 4)
sd0 <- 1
wts0 <- rep(1, n)

p <- length(beta0)
X <- cbind(rep(1, n), sapply(1:(p - 1), function(x) rnorm(n)))
mu <- X \%*\% beta0
y <- rnorm(n, mu, sd0 / sqrt(wts0))
dat <- cbind(y, X[, -1])
dat <- as.data.frame(dat)
names(dat) <- c("y", paste("x", 1:(p - 1), sep = ""))
nS <- 100
modular.x3 <- matrix(rnorm(length(dat$x3) * nS, mean = rep(dat$x3, nS), sd = 0.01),
                     nrow = length(dat$x3), ncol = nS)
mod1 <- modularLM(y ~ x1 + x2, post.var = list(x3 = modular.x3),
                  data = dat)
mod1.summary <- credibleInterval(mod1)
}
