% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleLM.R
\name{simpleLM}
\alias{simpleLM}
\title{Fit a Bayesian linear model}
\usage{
simpleLM(formula, data, wts, priors, n.samples = 1000, ...)
}
\arguments{
\item{formula}{a symbolic description of the regression model to be fit.}

\item{data}{An optional data frame containing the variables in the model.}

\item{wts}{A numeric vector of weights for the observations.}

\item{priors}{A list containing prior specifications for model parameters.}

\item{n.samples}{Number of posterior samples to draw.}

\item{...}{Additional arguments (not currently used).}
}
\value{
An object of class \code{fitLM} containing the fitted model and
posterior samples.
}
\description{
Fit a Bayesian linear model
}
