% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible.R
\name{credibleInterval}
\alias{credibleInterval}
\title{Generate credible intervals for model parameters}
\usage{
credibleInterval(mod.out, param, prob = 0.95)
}
\arguments{
\item{mod.out}{model output from running \code{simpleLM} or \code{modularLM}.}

\item{param}{(optional) A character vector of variable names for which
credible intervals are desired. If not specified, credible intervals for all
parameters are returned.}

\item{prob}{Specifies the high probability density region of the posterior
distribution. Default is 0.95, which corresponds to a 95\% credible interval.}
}
\value{
An object of class \code{credibleInterval} containing the lower and upper
bounds of the credible intervals for the specified parameters, along with
posterior medians.
}
\description{
Generate credible intervals for model parameters
}
