% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularLM.R
\name{modularLM}
\alias{modularLM}
\title{Fit a modular Bayesian linear model}
\usage{
modularLM(formula, data, post.var, wts, priors, ...)
}
\arguments{
\item{formula}{a symbolic description of the non-modular component of the
regression model to be fit.}

\item{data}{An optional data frame containing the variables in the model.}

\item{post.var}{A list tagged with modular predictor names, each containing
a matrix of posterior samples for that modular predictor. In case of
interaction terms involving a modular predictor, the interaction term should
also be included as a separate entry in this list.}

\item{wts}{A numeric vector of weights for the observations.}

\item{priors}{A list containing prior specifications for model parameters.}

\item{...}{Additional arguments (not currently used).}
}
\value{
An object of class \code{modularLM} containing the fitted model and
posterior samples.
}
\description{
Fit a modular Bayesian linear model
}
