% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleLM.R
\name{simpleLM}
\alias{simpleLM}
\title{Fit a Bayesian linear model}
\usage{
simpleLM(formula, data, wts, priors, n.samples = 1000, ...)
}
\arguments{
\item{formula}{a symbolic description of the regression model to be fit.}

\item{data}{An optional data frame containing the variables in the model.}

\item{wts}{A numeric vector of weights for the observations.}

\item{priors}{A list containing prior specifications for model parameters.}

\item{n.samples}{Number of posterior samples to draw.}

\item{...}{Additional arguments (not currently used).}
}
\value{
An object of class \code{fitLM} containing the fitted model and
posterior samples.
}
\description{
Fit a Bayesian linear model
}
\examples{
set.seed(1729)
n <- 100
beta0 <- c(5, -2, 1, 0)
sd0 <- 1
wts0 <- rep(1, n)

p <- length(beta0)
X <- cbind(rep(1, n), sapply(1:(p - 1), function(x) rnorm(n)))
mu <- X \%*\% beta0
y <- rnorm(n, mu, sd0 / sqrt(wts0))
dat <- cbind(y, X[, -1])
dat <- as.data.frame(dat)
names(dat) <- c("y", paste("x", 1:(p - 1), sep = ""))
mod1 <- lm(y ~ x1 * x2, data = dat, weights = wts0)
mod2 <- simpleLM(y ~ x1 * x2, data = dat)

mod1.summary <- summary(mod1)
mod2.summary <- credibleInterval(mod2)
print(mod2.summary)
}
